requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database & ORM
sqlmodel==0.0.14
sqlalchemy==2.0.25
alembic==1.12.1
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Job Queue
celery[redis]==5.3.4

# HTTP Client
httpx==0.25.2
aiohttp==3.12.15
requests==2.31.0

# Configuration
python-dotenv==1.0.0
pydantic-settings==2.1.0

# File handling
aiofiles==23.2.1
pillow==10.1.0

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0

# Monitoring & Logging
structlog==23.2.0

# Development tools
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1 